detroit_axle_qa_system_github_ready/python-backend/requirements.txt
fastapi
uvicorn[standard]
supabase
pydantic-settings

